requirements.txt
aws-cdk.core==1.91.0
aws-cdk.aws_s3==1.91.0
aws-cdk.aws-lambda==1.91.0
aws_cdk.aws_lambda_event_sources==1.91.0
aws-cdk.aws_dynamodb==1.91.0
aws-cdk.aws_apigateway==1.91.0

pytest==6.1.0
